+++ CHANGELOG.md
@@ -5,6 +5,9 @@ _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 
 ## [0.1.4] - INSERT MERGING DATE HERE
+### Added
+- Accept pending follow requests for private account
+
 ### Fixed
 - Only import instapy-chromedriver package when needed
 - Avoid user errors providing user names with caps